Implemented Chan's algorithm
This algorithm uses both Graham's Scan and the Jarvis March as
subroutines.  In implementing it, I needed to extend the Jarvis March
algorithm in a couple of ways.  First, it was necessary to make the
Jarvis March lazy, so that Chan's algorithm could abort it after a
specified number of steps.  Also, I needed to extend the next-pt
function to take a number of small precomputed convex hulls and
binary-search within those hulls for the best next point.